requirements/base.txt
argon2-cffi>=23.1,<26.0
Django>=5.2,<5.3
Pillow>=10.4,<12.0
celery[redis]>=5.4,<6.0
django-environ>=0.12,<1.0
django-htmx>=1.21,<2.0
django-phonenumber-field[phonenumberslite]>=8.0,<9.0
django-ratelimit>=4.1,<5.0
django-storages[s3]>=1.14,<2.0
django-unfold>=0.57,<1.0
djangorestframework>=3.15,<4.0
gunicorn>=23.0,<24.0
psycopg[binary]>=3.2,<4.0
redis>=5.0,<7.0
whitenoise>=6.8,<7.0
